+++ src/main/kotlin/org/jetbrains/changelog/ChangelogPluginConstants.kt
@@ -6,7 +6,7 @@ object ChangelogPluginConstants {
     const val GROUP_NAME = "changelog"
     const val EXTENSION_NAME = "changelog"
     const val PLUGIN_NAME = "Gradle Changelog Plugin"
-    const val MINIMAL_SUPPORTED_GRADLE_VERSION = "6.8"
+    const val MINIMAL_SUPPORTED_GRADLE_VERSION = "8.3"
 
     const val GET_CHANGELOG_TASK_NAME = "getChangelog"
     const val PATCH_CHANGELOG_TASK_NAME = "patchChangelog"