+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [2.5.0] - 2025-11-25
+
 ### Added
 
 - Add `headerInlineLink` property to preserve inline links in version headers instead of converting them to reference-style links JetBrains/gradle-changelog-plugin#274
@@ -21,7 +23,7 @@
 
 ### Added
 
-- Added the `--release-note-file` flag to `PatchChangelogTask` 
+- Added the `--release-note-file` flag to `PatchChangelogTask`
 
 ## [2.3.0] - 2025-07-26
 
@@ -328,7 +330,9 @@
 - `changelog` extension configuration
 - `getChangelog`/`patchChangelog` tasks
 
-[Unreleased]: https://github.com/JetBrains/gradle-changelog-plugin/compare/v2.3.0...HEAD
+[Unreleased]: https://github.com/JetBrains/gradle-changelog-plugin/compare/v2.5.0...HEAD
+[2.5.0]: https://github.com/JetBrains/gradle-changelog-plugin/compare/v2.4.0...v2.5.0
+[2.4.0]: https://github.com/JetBrains/gradle-changelog-plugin/compare/v2.3.0...v2.4.0
 [2.3.0]: https://github.com/JetBrains/gradle-changelog-plugin/compare/v2.2.1...v2.3.0
 [2.2.1]: https://github.com/JetBrains/gradle-changelog-plugin/compare/v2.2.0...v2.2.1
 [2.2.0]: https://github.com/JetBrains/gradle-changelog-plugin/compare/v2.1.2...v2.2.0